refactor: board 클래스 패키지 이동
game/board
