+++ src/main/java/chess/domain/game/Board.java
@@ -1,4 +1,4 @@
-package chess.domain.board;
+package chess.domain.game;
 
 import java.util.Arrays;
 import java.util.Collections;
