+++ src/main/java/chess/domain/game/ChessGame.java
@@ -1,6 +1,5 @@
 package chess.domain.game;
 
-import chess.domain.board.Board;
 import chess.domain.game.state.State;
 import chess.domain.piece.Color;
 import chess.domain.piece.Position;
