+++ src/main/java/chess/domain/game/state/Finished.java
@@ -1,6 +1,6 @@
 package chess.domain.game.state;
 
-import chess.domain.board.Board;
+import chess.domain.game.Board;
 import chess.domain.game.Score;
 import chess.domain.piece.Color;
 import chess.domain.piece.Position;
