+++ src/main/java/chess/domain/game/state/Ready.java
@@ -1,6 +1,6 @@
 package chess.domain.game.state;
 
-import chess.domain.board.Board;
+import chess.domain.game.Board;
 import chess.domain.game.Score;
 import chess.domain.game.Turn;
 import chess.domain.piece.Color;
