+++ src/test/java/chess/domain/game/state/FinishedTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.domain.board.Board;
+import chess.domain.game.Board;
 import chess.domain.piece.Position;
 
 public class FinishedTest {
